upgrading call alloy to avoid modelling-tasks build failure
modelling-tasks depends on a newer call-alloy version than this repository.
another solution is to downgrade the version used in modelling-tasks (which is the opposite of what one is supposed to do) and falling back on the version used here.
without doing this we run into a dependency conflict, unable to build modelling-tasks (due to a missing MonadIO class instance).
unfortunately i couldn't manage to implement a working solution, so i just error out.