fix PlantUML rendering failure
replace ForkOrJoin by Fork and Join respectively in the expected read input chart description for the unit test.
replace join with a fork in PlantUML validation file, as it looks like that outgoing from the fork node the states 2,1,1 and 2,2,3 are entered, and not joined.
there might have been a translation bug previously hiding in the PlantUML renderer, incorrectly determining the exact kind of ForkOrJoin node, or a regression the updated code introduced, as adjustments to posPlantUMLExamples were also necessary.