cart. product decomposition of combine node
using Cartesian-product to decompose InnerMostStates within SD regions bound to a CombineDiagram, along an attempt to classify connection types.
The crossed states can already be rewired, but Forks and Joins are still wip, as they will involve a bit more tracing in the background.
As of now function signatures are very noisy, but it should be possible to clean this up and unify the statements once the CD region can be lifted from a basic chart.
nodes are distinguish like done in the SD lift procedure; Left meaning the unaltered references and Right the ones belonging to crossed nodes (they infer labels of their atomic parts).