upper version bound relaxation to add modelling-tasks as direct module dependency
enabling unsafe upper bound version relaxation; allow-newer: true

otherwise, it is not possible to build this project with the modelling-tasks dependency addition explicitly used.

modelling-tasks was already added some time ago, but was not required to build any executable. However, direct inclusion is necessary as of now to express the instances for shuffling state charts.

this means accessing (modelling-tasks/src/Modelling/Auxiliary/Common.hs) on the module path.

the root cause of the relaxation, is latex-svg-image-0.2 that has a lower (upper version bound) to the base library, than what we can provide.

The reason to do this temp hack, is to avoid manually copying files from the modelling-tasks module path over to this project again, instead of accessing them directly when exposed.